Move classes to the root source folder
