+++ kategory/src/main/kotlin/kategory/Maps.kt
@@ -1,6 +1,4 @@
-package kategory.collections
-
-import kategory.Tuple2
+package kategory
 
 fun <K, V> mapOf(vararg tuple: Tuple2<K, V>): Map<K, V> =
         if (tuple.isNotEmpty()) tuple.map { it.a to it.b }.toMap() else emptyMap()
